+++ examples/webgpu/two-cubes/app.ts
@@ -81,8 +81,8 @@ const UNIFORM_BUFFER_SIZE = 4 * 16; // 4x4 matrix
 
 const CUBE_SHADER_LAYOUT: ShaderLayout = {
   attributes: [
-    {name: 'position', location: 0, format: 'float32x4'},
-    {name: 'uv', location: 1, format: 'float32x2'}
+    {name: 'position', location: 0, type: 'vec4<f32>'},
+    {name: 'uv', location: 1, type: 'vec2<f32>'}
   ],
   bindings: [
     {name: 'uniforms', location: 0, type: 'uniform'}
@@ -121,7 +121,7 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
       vs: SHADERS.vs,
       fs: SHADERS.fs,
       topology: 'triangle-list',
-      layout: CUBE_SHADER_LAYOUT,
+      shaderLayout: CUBE_SHADER_LAYOUT,
       attributes: {
         position: positionBuffer,
         uv: uvBuffer