+++ modules/core/src/adapter/types/accessor.ts
@@ -1,6 +1,6 @@
 // luma.gl, MIT license
 import type {Buffer} from '../resources/buffer';
-// import type {VertexFormat} from './vertex-formats';
+// import type {VertexFormat} from './data-formats';
 
 // ACCESSORS