+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # luma.gl CHANGELOG
 
+## v9.0.0-alpha.31
+
+- chore(core): Clean up buffer map
+- chore(core): default props cleanup (#1791)
+
 ## v9.0.0-alpha.30
 
 - fixe(core) boolean uniform fix (#1789)
@@ -9,7 +14,7 @@
 - docs: website examples (#1782)
 ## v9.0.0-alpha.29
 
-- feat(webgl): Implement ShaderLayout.bufferMap in webgl (#1780)
+- feat(webgl): Implement ShaderLayout.bufferLayout in webgl (#1780)
 - feat(engine): model cleanup (#1779)
 
 ## v9.0.0-alpha.28